refactor(date): move format date function to new file
