+++ components/DateSelector/DateSelector.tsx
@@ -1,14 +1,5 @@
 import { IconChevronLeft, IconChevronRight } from '@tabler/icons';
-
-export function formatDate(date: Date, options?: Intl.DateTimeFormatOptions) {
-  const dateOptions = options || {
-    weekday: 'short',
-    month: 'short',
-    day: 'numeric',
-  };
-
-  return date.toLocaleDateString('en-US', dateOptions);
-}
+import { formatDate } from './date';
 
 function DateSelector() {
   return (
